CI-159 - Code review fix up

+++ src/Sfa.Roatp.Register.Core/Services/IGetProviders.cs
@@ -6,6 +6,6 @@ namespace Sfa.Roatp.Register.Core.Services
     public interface IGetProviders
     {
         IEnumerable<RoatpProvider> GetAllProviders();
-        RoatpProvider GetProvider(long ukprn);
+        RoatpProvider GetProvider(int ukprn);
     }
 }
